requirements.txt
bcrypt==4.0.1
blinker==1.6.2
boto3==1.26.136
botocore==1.29.136
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
flask-paginate==2022.1.8
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
libretranslatepy==2.1.1
lxml==4.9.2
MarkupSafe==2.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.30.0
s3transfer==0.6.1
six==1.16.0
SQLAlchemy==2.0.14
translate==3.6.1
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.3.4
WTForms==3.0.1
zipp==3.15.0
